+++ changelog.md
@@ -14,6 +14,7 @@
 - `EmbedBase.__bool__` could return incorrect value.
 - `EmbedBase.__len__` could return incorrect value.
 - `EmbedBase.contents` excluded `.provider.name`
+- `Embed.copy_with` assigned `type` field with different name.
 
 ## 1.1.131 *\[2022-01-04\]*